feat(native.ex): using RustlerPrecompiled instead of Rustler module
Here we specify the url where the build artifacts will available for download (the actual building
and uploading is handled in CI). foce_build option is true if BLAKE3_BUILD env exists and value is
either  1 or true, this is helpful in library local development, where we can actually compile
locally with rust when needed.